Add files via upload
"Explore the seamless synergy of DevOps in action with this Dockerized 2048 Game project. Witness the harmonious orchestration of code and infrastructure as the classic game comes to life within Docker containers, ready for smooth deployment on AWS cloud. Dive into the world of continuous integration, automated testing, and efficient deployment pipelines."